:recycle: refactor notification handlers
